+++ molSim/chemical_datastructures/__init__.py
@@ -1,0 +1,2 @@
+from .molecule import Molecule
+from .molecule_set import MoleculeSet
