+++ molSim/ops/__init__.py
@@ -1,0 +1,3 @@
+from .descriptor import Descriptor
+from .clustering import Cluster
+from . import similarity_measures
