+++ molSim/tasks/__init__.py
@@ -1,0 +1,7 @@
+from .task import Task
+from .cluster_data import ClusterData
+from .compare_target_molecule import CompareTargetMolecule
+from .show_property_variation_with_similarity \
+        import ShowPropertyVariationWithSimilarity
+from .visualize_dataset import VisualizeDataset
+from .task_manager import TaskManager
